move all find_agents things into database
