+++ research_town/dbs/__init__.py
@@ -20,7 +20,7 @@ from .data import (
 from .db_log import LogDB
 from .db_paper import PaperDB
 from .db_progress import ProgressDB
-from .db_researcher import ResearcherDB
+from .db_researcher import AgentDB
 
 __all__ = [
     'IdeaDiscussionLog',
@@ -42,6 +42,6 @@ __all__ = [
     'Experiment',
     'LogDB',
     'PaperDB',
-    'ResearcherDB',
+    'AgentDB',
     'ProgressDB',
 ]
